+++ aten/src/ATen/Dispatch.h
@@ -4,7 +4,7 @@
 #include <c10/macros/Macros.h>
 #include <c10/util/Exception.h>
 #include <c10/util/Half.h>
-#include <c10/util/Metaprogramming.h>
+#include <torch/headeronly/util/Metaprogramming.h>
 #include <c10/util/complex.h>
 
 #ifdef __CUDACC__
