+++ aten/src/ATen/NestedTensorImpl.h
@@ -7,7 +7,7 @@
 #include <c10/core/TensorImpl.h>
 #include <c10/util/ArrayRef.h>
 #include <c10/util/Exception.h>
-#include <c10/util/Metaprogramming.h>
+#include <torch/headeronly/util/Metaprogramming.h>
 #include <c10/util/irange.h>
 
 namespace at::native {
