+++ aten/src/ATen/core/boxing/KernelFunction.h
@@ -4,7 +4,7 @@
 #include <ATen/core/boxing/BoxedKernel.h>
 #include <ATen/core/stack.h>
 #include <c10/core/DispatchKeySet.h>
-#include <c10/util/TypeList.h>
+#include <torch/headeronly/util/TypeList.h>
 #include <c10/util/intrusive_ptr.h>
 #include <atomic>
 #include <memory>
