+++ aten/src/ATen/core/boxing/impl/WrapFunctionIntoFunctor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <c10/core/CompileTimeFunctionPointer.h>
+#include <torch/headeronly/core/CompileTimeFunctionPointer.h>
 
 namespace c10::impl {
 namespace detail {
