+++ aten/src/ATen/core/boxing/impl/boxing.h
@@ -9,7 +9,7 @@
 
 #include <ATen/core/boxing/BoxedKernel.h>
 
-#include <c10/util/Metaprogramming.h>
+#include <torch/headeronly/util/Metaprogramming.h>
 #include <type_traits>
 
 namespace c10::impl {
