+++ aten/src/ATen/core/boxing/impl/make_boxed_from_unboxed_functor.h
@@ -4,8 +4,8 @@
 #include <ATen/core/boxing/OperatorKernel.h>
 #include <ATen/core/ivalue.h>
 #include <ATen/core/stack.h>
-#include <c10/util/Metaprogramming.h>
-#include <c10/util/TypeList.h>
+#include <torch/headeronly/util/Metaprogramming.h>
+#include <torch/headeronly/util/TypeList.h>
 #include <c10/util/intrusive_ptr.h>
 
 #include <utility>
