+++ aten/src/ATen/core/dispatch/OperatorEntry.h
@@ -7,7 +7,7 @@
 #include <c10/core/DispatchKey.h>
 #include <c10/core/PyHandleCache.h>
 #include <c10/core/SafePyObject.h>
-#include <c10/util/Metaprogramming.h>
+#include <torch/headeronly/util/Metaprogramming.h>
 #include <c10/util/flat_hash_map.h>
 
 #include <ATen/core/dispatch/CppSignature.h>
