+++ c10/test/util/TypeIndex_test.cpp
@@ -1,4 +1,4 @@
-#include <c10/util/Metaprogramming.h>
+#include <torch/headeronly/util/Metaprogramming.h>
 #include <c10/util/TypeIndex.h>
 #include <gtest/gtest.h>
