+++ c10/util/TypeTraits.cpp
@@ -1,1 +1,1 @@
-#include <c10/util/TypeTraits.h>
+#include <torch/headeronly/util/TypeTraits.h>
