+++ torch/custom_class.h
@@ -7,9 +7,9 @@
 #include <ATen/core/op_registration/infer_schema.h>
 #include <ATen/core/stack.h>
 #include <c10/util/C++17.h>
-#include <c10/util/Metaprogramming.h>
-#include <c10/util/TypeList.h>
-#include <c10/util/TypeTraits.h>
+#include <torch/headeronly/util/Metaprogramming.h>
+#include <torch/headeronly/util/TypeList.h>
+#include <torch/headeronly/util/TypeTraits.h>
 #include <torch/custom_class_detail.h>
 #include <torch/library.h>
